refactor: Move theme to the geti/ui/theme
