+++ web_ui/packages/ui/theme/geti/geti-light.module.css
