+++ web_ui/src/assets/index.scss
@@ -1,6 +1,6 @@
-@import '../theme/geti/geti-global.module.css';
-@import '../theme/geti/geti-dark.module.css';
-@import '../theme/geti/geti-medium.module.css';
+@import '../../packages/ui/theme/geti/geti-global.module.css';
+@import '../../packages/ui/theme/geti/geti-dark.module.css';
+@import '../../packages/ui/theme/geti/geti-medium.module.css';
 @import 'react-notifications-component/dist/theme.css';
 @import 'fonts/intelone/intelone.css';
