+++ web_ui/src/pages/annotator/components/footer/fps.component.tsx
@@ -4,8 +4,8 @@
 import { dimensionValue, useMediaQuery, useNumberFormatter, View } from '@geti/ui';
 import { Fps } from '@geti/ui/icons';
 
+import { biggerThanQuery } from '../../../../../packages/ui/theme/queries';
 import { Video, VideoFrame } from '../../../../core/media/video.interface';
-import { biggerThanQuery } from '../../../../theme/queries';
 
 export const FPS = ({ mediaItem, className }: { mediaItem: Video | VideoFrame; className?: string }): JSX.Element => {
     const formatter = useNumberFormatter({
