+++ web_ui/src/pages/annotator/components/footer/project-name.component.tsx
@@ -3,8 +3,8 @@
 
 import { useMediaQuery, View } from '@geti/ui';
 
+import { isLargeSizeQuery } from '../../../../../packages/ui/theme/queries';
 import { TruncatedTextWithTooltip } from '../../../../shared/components/truncated-text/truncated-text.component';
-import { isLargeSizeQuery } from '../../../../theme/queries';
 
 import classes from './annotator-footer.module.scss';
