+++ web_ui/src/pages/annotator/components/footer/training-progress/training-progress.component.tsx
@@ -3,9 +3,9 @@
 
 import { Flex, Text, useMediaQuery, useNumberFormatter, View } from '@geti/ui';
 
+import { isLargeSizeQuery, isMediumLargeSizeQuery } from '../../../../../../packages/ui/theme/queries';
 import { ProgressBar } from '../../../../../shared/components/progress-bar/progress-bar.component';
 import { trimText } from '../../../../../shared/utils';
-import { isLargeSizeQuery, isMediumLargeSizeQuery } from '../../../../../theme/queries';
 
 import classes from './training-progress.module.scss';
