+++ web_ui/src/pages/annotator/components/navigation-toolbar/navigation-breadcrumbs/tasks-breadcrumbs.component.tsx
@@ -3,9 +3,9 @@
 
 import { Item, Picker, useMediaQuery } from '@geti/ui';
 
+import { isLargeSizeQuery } from '../../../../../../packages/ui/theme/queries';
 import { Task } from '../../../../../core/projects/task.interface';
 import { hasEqualId } from '../../../../../shared/utils';
-import { isLargeSizeQuery } from '../../../../../theme/queries';
 import { useIsSceneBusy } from '../../../hooks/use-annotator-scene-interaction-state.hook';
 import { usePrediction } from '../../../providers/prediction-provider/prediction-provider.component';
 import { BreadcrumbSegment } from './breadcrumb-segment.component';
