+++ web_ui/src/pages/annotator/components/secondary-toolbar/secondary-toolbar.component.tsx
@@ -6,12 +6,12 @@ import { AnimatePresence, motion } from 'framer-motion';
 import { isEmpty, negate, noop } from 'lodash-es';
 import { useSearchParams } from 'react-router-dom';
 
+import { isLargeSizeQuery } from '../../../../../packages/ui/theme/queries';
 import { getAnnotationStateForTask } from '../../../../core/annotations/utils';
 import { MEDIA_ANNOTATION_STATUS } from '../../../../core/media/base.interface';
 import { MediaItem } from '../../../../core/media/media.interface';
 import { isClassificationDomain } from '../../../../core/projects/domains';
 import { ANIMATION_PARAMETERS } from '../../../../shared/animation-parameters/animation-parameters';
-import { isLargeSizeQuery } from '../../../../theme/queries';
 import { useProject } from '../../../project-details/providers/project-provider/project-provider.component';
 import { AnnotationToolContext, ANNOTATOR_MODE } from '../../core/annotation-tool-context.interface';
 import { useAnnotatorMode } from '../../hooks/use-annotator-mode';
