+++ web_ui/src/pages/annotator/providers/annotator-context-menu-provider/context-menu/context-menu.test.tsx
@@ -1,9 +1,9 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
+import { ThemeProvider } from '@geti/ui/theme';
 import { fireEvent, render, screen } from '@testing-library/react';
 
-import { ThemeProvider } from '../../../../../theme/theme-provider.component';
 import { ContextMenu } from './context-menu.component';
 
 describe('ContextMenu', () => {
