+++ web_ui/src/pages/annotator/tools/watershed-tool/secondary-toolbar.component.tsx
@@ -6,8 +6,8 @@ import { Key, useEffect } from 'react';
 import { Divider, Flex, Text, useMediaQuery } from '@geti/ui';
 import { isEmpty } from 'lodash-es';
 
+import { isLargeSizeQuery } from '../../../../../packages/ui/theme/queries';
 import { filterOutExclusiveLabel } from '../../../../core/labels/utils';
-import { isLargeSizeQuery } from '../../../../theme/queries';
 import { AcceptRejectButtonGroup } from '../../components/accept-reject-button-group/accept-reject-button-group.component';
 import { ToolType } from '../../core/annotation-tool-context.interface';
 import { useAnnotationScene } from '../../providers/annotation-scene-provider/annotation-scene-provider.component';
