+++ web_ui/src/pages/create-project/components/select-project-template/single-task-template.component.tsx
@@ -5,9 +5,9 @@ import { useEffect, useState } from 'react';
 
 import { Grid, minmax, repeat, useMediaQuery } from '@geti/ui';
 
+import { isLargeSizeQuery } from '../../../../../packages/ui/theme/queries';
 import { useFeatureFlags } from '../../../../core/feature-flags/hooks/use-feature-flags.hook';
 import { SUBDOMAIN } from '../../../../core/projects/project.interface';
-import { isLargeSizeQuery } from '../../../../theme/queries';
 import { Card } from '../card.component';
 import { DomainCardsMetadata, SingleTemplateProps } from './project-template.interface';
