+++ web_ui/src/pages/create-project/components/select-project-template/task-chain-template.component.tsx
@@ -6,10 +6,10 @@ import { useEffect } from 'react';
 import { Flex, Heading, Radio, RadioGroup, Text, useMediaQuery, View } from '@geti/ui';
 import { isEqual } from 'lodash-es';
 
+import { isLargeSizeQuery } from '../../../../../packages/ui/theme/queries';
 import { DOMAIN } from '../../../../core/projects/core.interface';
 import { Arrow } from '../../../../shared/components/arrow/arrow.component';
 import { idMatchingFormat } from '../../../../test-utils/id-utils';
-import { isLargeSizeQuery } from '../../../../theme/queries';
 import { Card } from '../card.component';
 import { TaskChainMetadata, TaskChainTemplateProps } from './project-template.interface';
