+++ web_ui/src/pages/create-project/new-project-dialog.component.tsx
@@ -22,12 +22,12 @@ import { OverlayTriggerState } from '@react-stately/overlays';
 import { isEmpty } from 'lodash-es';
 import { useNavigate } from 'react-router-dom';
 
+import { isLargeSizeQuery } from '../../../packages/ui/theme/queries';
 import { DOMAIN } from '../../core/projects/core.interface';
 import { useProjectActions } from '../../core/projects/hooks/use-project-actions.hook';
 import { paths } from '../../core/services/routes';
 import { useWorkspaceIdentifier } from '../../providers/workspaces-provider/use-workspace-identifier.hook';
 import { TooltipWithDisableButton } from '../../shared/components/custom-tooltip/tooltip-with-disable-button';
-import { isLargeSizeQuery } from '../../theme/queries';
 import { InfoSection } from './components/info-section/info-section.component';
 import { getStepInfo, getTotalSteps } from './components/utils';
 import { CreateProjectButton } from './create-project-button/create-project-button.component';
