+++ web_ui/src/pages/landing-page/landing-page-sidebar/landing-page-sidebar.component.tsx
@@ -5,6 +5,7 @@ import { Divider, Flex, Link as SpectrumLink, useMediaQuery, View } from '@geti/
 import { InfoOutline, Policy, ProjectsIcon, UserIcon } from '@geti/ui/icons';
 import { Link } from 'react-router-dom';
 
+import { isLargeSizeQuery } from '../../../../packages/ui/theme/queries';
 import { useFeatureFlags } from '../../../core/feature-flags/hooks/use-feature-flags.hook';
 import { paths } from '../../../core/services/routes';
 import { useFirstWorkspaceIdentifier } from '../../../providers/workspaces-provider/use-first-workspace-identifier.hook';
@@ -13,7 +14,6 @@ import { MenuOption } from '../../../shared/components/menu-option.interface';
 import { ShowForOnPrem } from '../../../shared/components/show-for-onprem/show-for-onprem.component';
 import { SidebarMenu } from '../../../shared/components/sidebar-menu/sidebar-menu.component';
 import { idMatchingFormat } from '../../../test-utils/id-utils';
-import { isLargeSizeQuery } from '../../../theme/queries';
 import { OrganizationsPicker } from './organizations-picker/organizations-picker.component';
 import { StorageUsage } from './storage-usage/storage-usage.component';
