+++ web_ui/src/shared/components/configurable-parameters/cp-item/cp-editable-item/cp-number/cp-number.component.tsx
@@ -5,7 +5,7 @@ import { useMemo } from 'react';
 
 import { Flex, NumberField, Slider, Text, useMediaQuery, useNumberFormatter } from '@geti/ui';
 
-import { isLargeSizeQuery } from '../../../../../../theme/queries';
+import { isLargeSizeQuery } from '../../../../../../../packages/ui/theme/queries';
 import { NumberGroupParams } from '../../../configurable-parameters.interface';
 import { ResetButtonHandler } from '../cp-editable-item.interface';
