+++ web_ui/src/shared/components/dataset-import-dnd/dataset-import-dnd.component.tsx
@@ -8,10 +8,10 @@ import { DatasetImport as DatasetImportIcon } from '@geti/ui/icons';
 import { Link } from 'react-router-dom';
 import { useMediaQuery } from 'usehooks-ts';
 
+import { isExtraLargeSizeQuery } from '../../../../packages/ui/theme/queries';
 import { IMPORT_DATASET_LEARN_MORE } from '../../../core/const';
 import { NOTIFICATION_TYPE } from '../../../notification/notification-toast/notification-type.enum';
 import { useNotification } from '../../../notification/notification.component';
-import { isExtraLargeSizeQuery } from '../../../theme/queries';
 import { DropZone, onDropFiles } from '../../drag-and-drop/drag-and-drop.component';
 import { onValidFileList } from '../../utils';
